Refactor: 판매자센터 페이지 통합
Refs: #7
